Partial update better
